created manager folder, exported manager activity code + update comment count function
