+++ server/controllers/activity.ts
@@ -8,11 +8,11 @@ import type { GameData, GamesData } from '../../types/game.type';
 import type { StoriesData, StoryElement } from '../../types/story.type';
 import { ImageType } from '../../types/story.type';
 import { Activity } from '../entities/activity';
-import { Comment } from '../entities/comment';
 import { Game } from '../entities/game';
 import { Image } from '../entities/image';
 import { UserType } from '../entities/user';
 import { VillagePhase } from '../entities/village';
+import { getActivities, getActivitiesCommentCount } from '../manager/activity';
 import { AppError, ErrorCode } from '../middlewares/handleErrors';
 import { getQueryString } from '../utils';
 import { AppDataSource } from '../utils/data-source';
@@ -21,153 +21,6 @@ import { commentController } from './comment';
 import { Controller } from './controller';
 
 const activityController = new Controller('/activities');
-
-const getActivitiesCommentCount = async (ids: number[]): Promise<{ [key: number]: number }> => {
-  if (ids.length === 0) {
-    return {};
-  }
-  const queryBuilder = await AppDataSource.getRepository(Activity)
-    .createQueryBuilder('activity')
-    .select('activity.id')
-    .addSelect('IFNULL(`commentCount`, 0) + IFNULL(`activityCount`, 0)', 'comments')
-    .leftJoin(
-      (qb) => {
-        qb.select('comment.activityId', 'cid').addSelect('COUNT(comment.id)', 'commentCount').from(Comment, 'comment').groupBy('comment.activityId');
-        return qb;
-      },
-      'comments',
-      '`comments`.`cid` = activity.id',
-    )
-    .leftJoin(
-      (qb) => {
-        qb.select('activity.responseActivityId', 'aid')
-          .addSelect('COUNT(activity.id)', 'activityCount')
-          .from(Activity, 'activity')
-          .where('activity.status != :status', { status: `${ActivityStatus.DRAFT}` })
-          .groupBy('activity.responseActivityId');
-        return qb;
-      },
-      'activities',
-      '`activities`.`aid` = activity.id',
-    )
-    .where('activity.id in (:ids)', { ids })
-    .getRawMany();
-  return queryBuilder.reduce((acc, row) => {
-    acc[row.activity_id] = parseInt(row.comments, 10) || 0;
-    return acc;
-  }, {});
-};
-
-type ActivityGetter = {
-  limit?: number;
-  page?: number;
-  phase?: number | null;
-  villageId?: number;
-  type?: string[];
-  subType?: number | null;
-  countries?: string[];
-  pelico?: boolean;
-  userId?: number;
-  status?: number;
-  responseActivityId?: number;
-  delayedDays?: number;
-  hasVisibilitySetToClass?: boolean;
-  teacherId?: number;
-  visibleToParent: boolean;
-};
-
-const getActivities = async ({
-  limit = 200,
-  page = 0,
-  villageId,
-  type = [],
-  subType = null,
-  countries,
-  phase = null,
-  pelico = true,
-  status = 0,
-  userId,
-  responseActivityId,
-  delayedDays,
-  hasVisibilitySetToClass,
-  teacherId,
-  visibleToParent,
-}: ActivityGetter) => {
-  // get ids
-  let subQueryBuilder = AppDataSource.getRepository(Activity).createQueryBuilder('activity').where('activity.status = :status', { status });
-  if (villageId !== undefined) {
-    subQueryBuilder = subQueryBuilder.andWhere('activity.villageId = :villageId', { villageId });
-  }
-  if (type.length > 0) {
-    subQueryBuilder = subQueryBuilder.andWhere('activity.type IN (:type)', { type });
-  }
-  if (subType !== null) {
-    subQueryBuilder = subQueryBuilder.andWhere('activity.subType = :subType', { subType });
-  }
-  if (phase !== null) {
-    subQueryBuilder = subQueryBuilder.andWhere('activity.phase = :phase', { phase });
-  }
-  if (delayedDays !== undefined) {
-    // * Memo: we only select activity with updateDate + delayedDays lesser than current date
-    subQueryBuilder = subQueryBuilder.andWhere(`DATE_ADD(activity.updateDate, INTERVAL :delayedDays DAY) <= CURDATE()`, { delayedDays: delayedDays });
-  }
-  if (visibleToParent) {
-    // * Here if user is a parent, we only select activities with the attribute is set to true
-    subQueryBuilder = subQueryBuilder.andWhere('activity.isVisibleToParent = :visibleToParent', { visibleToParent });
-  }
-  if (hasVisibilitySetToClass !== undefined && teacherId !== undefined) {
-    subQueryBuilder = subQueryBuilder.andWhere('activity.user = :teacherId', { teacherId: teacherId });
-  }
-
-  if (responseActivityId !== undefined) {
-    subQueryBuilder = subQueryBuilder.andWhere('activity.responseActivityId = :responseActivityId', { responseActivityId });
-  } else if (userId !== undefined) {
-    subQueryBuilder = subQueryBuilder.innerJoin('activity.user', 'user').andWhere('user.id = :userId', {
-      userId,
-    });
-  } else if (pelico && countries !== undefined && countries.length > 0) {
-    subQueryBuilder = subQueryBuilder
-      .innerJoin('activity.user', 'user')
-      .andWhere('((user.countryCode IN (:countries) AND user.type >= :userType) OR user.type <= :userType2)', {
-        countries,
-        userType: UserType.TEACHER,
-        userType2: UserType.MEDIATOR,
-      });
-  } else if (pelico && countries !== undefined && countries.length === 0) {
-    subQueryBuilder = subQueryBuilder.innerJoin('activity.user', 'user').andWhere('user.type <= :userType2', {
-      userType2: UserType.MEDIATOR,
-    });
-  } else if (!pelico && countries !== undefined && countries.length > 0) {
-    subQueryBuilder = subQueryBuilder.innerJoin('activity.user', 'user').andWhere('user.countryCode IN (:countries) AND user.type >= :userType', {
-      countries,
-      userType: UserType.TEACHER,
-    });
-  } else if (!pelico && countries !== undefined) {
-    return [];
-  }
-
-  const activities = await subQueryBuilder
-    .orderBy('activity.isPinned', 'DESC')
-    .addOrderBy('activity.updateDate', 'DESC')
-    .limit(limit)
-    .offset(page * limit)
-    .getMany();
-
-  const ids = activities.map((a) => a.id);
-  if (ids.length === 0) {
-    return [];
-  }
-
-  const comments = await getActivitiesCommentCount(ids);
-  for (const activity of activities) {
-    if (comments[activity.id] !== undefined) {
-      activity.commentCount = comments[activity.id];
-    } else {
-      activity.commentCount = 0;
-    }
-  }
-  return activities;
-};
 
 // --- Get all activities. ---
 activityController.get({ path: '', userType: UserType.OBSERVATOR }, async (req: Request, res: Response) => {
